Move usePreactSignalStore to it's own file
